Move vertx tests into the right directory
