+++ simpleclient_vertx/src/test/java/io/prometheus/client/vertx/MetricsHandlerTest.java
@@ -1,4 +1,4 @@
-package io.prometheus.client.exporter;
+package io.prometheus.client.vertx;
 
 import io.prometheus.client.CollectorRegistry;
 import io.prometheus.client.Gauge;
